audio_drift_detection/requirements.txt
librosa>=0.10.0
numpy>=1.24.0
scipy>=1.10.0
torch>=2.0.0
scikit-learn>=1.2.0
river>=0.21.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
pyyaml>=6.0
soundata>=0.1.0
soundfile>=0.12.0
